Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/workers/semantics/multiple-workers/003-expected.txt

CONSOLE MESSAGE: line 3: ReferenceError: Can't find variable: Worker

Harness Error (FAIL), message = ReferenceError: Can't find variable: Worker

TIMEOUT creating 3 nested dedicated workers Test timed out
